requirements.txt
django==4.2
python-decouple==3.8
psycopg2-binary==2.9.10
pre-commit>=3.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pillow==12.0.0
django-unfold==0.68.0
celery==5.5.3
redis==7.0.1
googletrans==4.0.0-rc1
django-modeltranslation
Faker==38.2.0
langdetect==1.0.9
Unidecode==1.4.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
